OP3-Demo: Adds basis for adding an image saving message
Adds .srv for the SaveImage service that can be listened to, with the
end goal of being able to interact from the web interface.

Adds the .msg with any parameters to send (a bit of a dummy in terms of
content, will add more content as the implementation fleshes out).

Modifies the CMakeLists.txt to include the new files.

Includes generated SaveImage.h in ball_detector.h, adds prototype for
image saving callback.

Signed-off-by: Nicholas Gilbert <gnick@pdx.edu>